Add exception checking to script. Replace graph file.
Turns out NetworkX didn't like the GEXF file I had, since the node ID became the node label. I switched that out for the Pajek NET file I had, which imports the ID as is. Also, print out the row number that an exception occurs on in the data file.