Remove GeoJSON data. Update network files, README, license.
The GeoJSON data is only used in my Shiny app, so I removed it from this
repository. Same goes for the abbreviated network. Update the README for
the network files to be more descriptive. Update the main README to
clarify a couple of points. Update the license file to change the year.